Couple of bug fixes to buffer.x and refactor to be more maintainable.
buffer.x is allowed to go up to the number of columns due to the way
that the non-autowrapping mode works - this is taken care of in
handlePrint.  But other code has been using buffer.x without taking
this into consideration, so two things are done here: one, the
clamping for reading is now done on the property, and two the one
scenario that I found was not doing it has been covered.

Style-wise, I replaced some of the hex constants with the ascii
comment like: 0x3f /* ? */ with the more readable (and searchable when
debugging) UInt8 (ascii: "?").