Terminal emulation fixes for alt-screen and attributes.
The sequence "1049" in set mode is an alias for "1048" and "1047" and
should be clear the alt buffer when going to the normal buffer on that
scenario.  This was causing "junk" to remain on the alt-buffer because
it was not being cleared for certain TERM definitions (on the iOS
client I was setting TERM to xterm-256color, while on the Mac I was
using xterm-color, so it looked like an iOS bug, but it was a general
emulation bug).

The second problem is that I missunderstood how "OptionSet.remove" works,
the operation itself is mutating, so

style.remove (.underline)

Would be enough to remove the .underline from the optionset, but instead,
I was doing:

style = style.remove (.underline)

And this operation instead return the intersection of the existing value
and the passed value, meaning that I was never clearing the state when
precise commands were used, only when a full reset was issued (which was
the case with xterm-color)