New strategy to implement UITextInput and family.
The previous code never managed to get any strings to be sent during
insertion, and I am not quite sure what caused this.   The new approach
rather than trying to map the terminal dimensions to the screen, instead
uses an offline char array called 'storage' that is used during things like
the dictation and where both the UITextInput selection and marked text
are applied - which is independent of the "selection" SelectionService field
in the TerminalView itself.

This makes dicatation work.

Currently this seems to regress international input (in particular the
Chinese input) but I am not sure if the reason is the accessorView and
AutoLayout (I do get a warning error about a conflict), or something
else (setting autoresizemask on the inputAccessoryView did not seem to
fix this).

Some hard earned lessons:

- insertText (text: String) sometimes get a bunch of spaces, but those
  are ok, because we need to track
  _markedTextRange/_selectedTextRange, and remove the text from the
  storage before applying the text.

  This solves the mysterious 10 white spaces, followed by the dictated content.

- The new logic that keeps storage elsewhere, sadly, has the
  disadvantage of not sending chiense input to the target.  So this
  approach regresses this part (this is handled by the setMarkText
  code).

- the logic in insertText currently sets sendData to either the text in storage,
  or the text being sent, and this is not right, and regresses some basic functionality.

Seems like I made a bunch of progress in dictation, but regressed the
input methods.