Sixel progress: I do like this approach a lot more.  Now, when an image has bee completed, it is the job of the front-end to insert it in a way that is suitable.
In the Mac+iOS cases, I divide the image in bands, and attach each
band to the display, this ensures that images work when scrolling
and that no odd repaint side effect happens when we update portions
that were not the "reference" row that contained the image.   Now,
every row has the image it needs to display.

Leveraged the Bufferline structure, and placed an optional array of
images there.   The images contain the column where they were
attached, so the front-end can iterate over the images (typically,
no more than one) to redraw.

Fixed also the HLS to RGB conversion, and does this without requiring
the core of the terminal to rely on UIKit or AppKit, so this should
be finally back to reusable on non-Apple platforms.