Rename createImage to createImageFromBitmap.
Introduce createImage that is a more general purpose facility that
expects the front-end to decode the image (could be a PNG or JPEG
file), and that takes a width and height specification that contain
a payload.

On the Apple View implementation, this takes care of scaling based on
the iTerm2 rules the image, based on these inputs, and based on the
PreserveRatio attribute.

The Terminal delegate method iTermContent is no longer a general
purpose handler, but instead merely a fallback that receives the data
if it was not handled before - this means that the image is
pre-processed during the "1337" iTerm handler, and is not delivered to
the iTerm2 custom handler.   The parsing is now done centraly, and a
UI should implement the createImage method.