iOS TerminalView: change the way selection works in a few ways.  Now, it behaves more like a text input control in iOS.  A long-press will add a "Select" option.
New selection markers are rendered, which can be used to adjust the
selection and these will scroll the view as well if they are near the
edge.

fuzzyHasSelectionBackground: Fixes the selection rendering on iOS, the
logic to return always true was wrong, it has to be true in one case,
false in the other.

Some behavioral changes to better match the iOS behavior:

- Long press will open a context menu, where it is possible to start a
  selection.

- Selecting "Select" after a long press, shows a new menu with
  additional context options.

- Menu is shown at the end of a pan operation when selection is on.

Additional property to control the color handles for the selection