Implement the helper keyboard.
SwiftTerm already has an inputAccessoryView that would provide a
handful of terminal commands that the iOS keyboard does not provide,
but lacked additional keys.

Now, it features a keyboard that can be displayed on demand that contains
helper keys, like function keys and other common keys.

Additionally, I introduced a darker color for some keys that do not
insert text, but have more editing capabilities, and it works on both
light and dark modes, and supports switching dynamically.