This shows a possible implementation path to draw the carret during the buffer update.
It does not attempt to be efficient, so every cursor update, triggers a full
screen refresh.

It also does not pick up the color from the terminal session, and in fact,
this highlights that we need to add support for tracking both foreground and
background colors in the terminal.

Support for other shapes than block is not supported, this would merely change
the color for the character, nor are any of the other shapes supported with
this approach.